+++ K8sUtils/private/Classes.ps1
@@ -14,6 +14,7 @@ When in doubt, restart the session.
 enum Status {
     Unknown
     Running
+    Timeout
     Crash
     ConfigError
 }
@@ -38,7 +39,7 @@ function mapContainerStatus($containerStatus) {
         return $containerStatus.state.waiting.reason -eq "CrashLoopBackOff" ? [Status]::Crash : [Status]::ConfigError,($containerStatus.state.waiting.reason)
     }
     if ((Get-Member -InputObject $containerStatus.state -Name 'terminated') -and $containerStatus.state.terminated) {
-        return [Status]::Crash,($containerStatus.state.terminated.reason)
+        return $containerStatus.state.terminated.reason -eq 'completed' ? [Status]::Running : [Status]::Crash,($containerStatus.state.terminated.reason)
     }
     return [Status]::Unknown,"Possible timeout or probe failure"
 }